MemoryPressure: tune memory pressure settings for rapid mem usage changes
When playing an asset with video, frequent seek/jump operations within
a short period may cause rapid memory increases. The current pressure
settings may not allow (enough) memory release in time to avoid an
app running in a container to be killed due it reaching the memory
limits. To allow sufficient, in time, memory release, the release
needs to happen with the "synchrounous" flag set to true, allowing
full garbage collection cycle when reaching critical limits.

Furthermore, the critical threshold needs to be lowered as well
considering that the release is not instantaneous and on embedded
devices the 95% original threshold does not allow enough room for
mem release on apps with lower allowed memory usage limits.