[MediaStream][GStreamer] Hardware decoder/sinks do not have src pads https://bugs.webkit.org/show_bug.cgi?id=276170
Reviewed by Philippe Normand.

In the MediaStream use case, we are trying to get decoder stats from a pad probe at the src pad of the video decoder.
The problem comes with the decoder/sink elements that don't have one because they are the end of the pipeline.

* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::configureVideoDecoder):

Canonical link: https://commits.webkit.org/280622@main