Add TTSClient based speech synthesizer
This commit adds an implementation of PlatformSpeechSynthesizer
using TTS Client. It converts text to audio with Thunder TTS plugin
or Firebolt and plays it with GStreamer. Disabled in compile time
with USE_TTS_CLIENT option.

GST + Flite impl was hidden under USE_FLITE option. It is default
choice still for SPEECH_SYNTHESIS is enabled.