Support whitelisting URLs for local universal access
When launching the browser with a local (file) URL, the page may
need to load scripts that reside on a different volume.
The commit 448487f4e635fac0719e05793ab6fba164ae361a introduced a
change to use the posix version of FileSystem::getFileDeviceId()
instead of the glib one. This affected the return value which is
now non-zero, fixing a bug, but changing the behavior of
FileSystem::filesHaveSameVolume() compared to older versions and
consequently the behavior of SecurityOrigin::canDisplay() when a
page loads a script that resides on a different volume.
While this can be overcome by enabling the setting to allow
universal access from file urls using the API
webkit_settings_set_allow_universal_access_from_file_urls(),
that will allow a wider access than needed, as the access is
only required from a limited number of trusted local files.
This new API introduces a way to overcome this issue and allow
only the access that is required.