Fix content type frame rate limit
Decoder limits are set by a build configuration in the form of
widthxheight@framerate. However, framerate limit is only considered
when the width and height are the ones in the build config. If the
actual requested resolution is different, no framerate limit is
ever considered.
This change brings back 2.38 behavior where limits where considered
for each variable independently. While not ideal, as lower
resolutions might support higher frame rates, and the way the build
config ties semantically the framerate to the resolution specified,
an alternative fix would require either an absolute max rate to be
specified, or possibly different sets of resolution + framerate
limits, which might be overkill to this purpose.