RDKDEV-1148: Release CDM resources when not used
When HTMLMediaElement has been created in WPE and HTMLMediaElement::setMediaKeys()
method was called to set DRM keys, CDMThunder implementation already allocates CDM
resources and they're not released until HTMLMediaElement object is destructed.
HTMLMediaElement destruction is non-deterministic and it's relying on Garbage Collector.
In this case very often we observe that after WPE exit (WPE process is not closed,
but some #boot address is being loaded), CDM resources are not released immediately
and CDM can't be used by other applications like IP Player or Netflix.

Proposed solution introduces following changes:

1) CDMPrivateThunder implementation doesn't occupy CDM resources permanently,
as it only needs CDM resources to check if server certificate is supported
so CDM resources can be acquired before check and released afterwards. The CDM
resources are reallocated again if any code asks for the OpenCDMSystem instance
again.

2) CDMInstanceThunder object's CDM resources are released on HTMLMediaElement::stop()
to not wait for Garbage Collector

See: https://github.com/WebPlatformForEmbedded/WPEWebKit/pull/1580

Co-authored by: Enrique Ocaña González <eocanha@igalia.com>

WARNING: While this commit is currently working for all the tested use cases,
including reassigning one MediaKeys from one <video> to a different one, it's
generally a bad idea to mess with the lifecycle of objects in this way. There
can be some other unforeseen usage patterns that may create issues in the
future. The proper solution to the problem would be to ensure that when the
WebKitBrowser Thunder plugin is destroyed, all the JavaScript objects held by
the WebProcess are properly garbage collected and therefore destroyed. This
would release the OpenCDMSystem and other valuable external resources without
requiring the preventive release hack done in this commit.