[MSE][GStreamer] Early reset m_isEndReached (EOS) when about to seek
When a video has reached End Of Stream and a (backwards) seek is done,
the media player won't ever complete the seek because the EOS condition
is still in place and no buffers are ever going to be pushed. A
successful seek completion would reset the EOS condition, but that won't
happen since the seek won't even start.

This change resets m_isEndReached as soon as possible (on
willSeekToTarget()), so the seek can happen.

See: https://github.com/WebPlatformForEmbedded/WPEWebKit/pull/1568#issuecomment-3448502697